Add orders/inventory modules & refactor entities
Introduce new orders and inventory layers and refactor related entities. Highlights:
- Add controllers, services and repositories for orders (Deliveries, LoyaltyPoints, OrderItems, Orders, PaymentMethods, Subscriptions, TrayReturns) and inventory movements.
- Add DTOs for inventory and orders payloads (InventoryMovementDTO, CheckoutDTO, OrderItemDTO, LoyaltyPointDTO, etc.).
- Implement InventoryMovementsService with inventory quantity updates; extend InventoryService with updateQuantity and productVariant linking.
- Add InventoryMovementsRepository and multiple new JPA repositories for orders and trays.
- Refactor entities: use BigDecimal for monetary fields (price, loyaltyDiscount), add relations (OrderItem, Batch, TrayReturn, Order, profiles), set default timestamps/flags, rename LoyaltyPoints -> LoyaltyPoint, and adjust column names/nullable constraints.
- Extend enums with new values (InvMovementType, PaymentType, StatusDeliveryType, StatusOrder).
- Minor DTO and payload adjustments (InventoryDTO adds variant, ProductCatalogDTO price -> BigDecimal).

These changes add core CRUD/DTO/repository scaffolding for order and inventory flows and standardize types/relations across entities.

+++ src/main/java/giuliacrepaldi/Lanzi_Orto_Urbano_Management/entities/orders/OrderItem.java
@@ -5,6 +5,7 @@ import giuliacrepaldi.Lanzi_Orto_Urbano_Management.entities.products.ProductVari
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Getter
@@ -27,7 +28,7 @@ public class OrderItem {
     private Integer quantity;
 
     @Column(nullable = false)
-    private Double price;
+    private BigDecimal price;
 
     @ManyToOne
     @JoinColumn(name = "order_id", nullable = false)
@@ -41,4 +42,6 @@ public class OrderItem {
     @ManyToOne
     @JoinColumn(name = "batch_id", nullable = false)
     private Batch batch;
+
+
 }
